api/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.104.0
psycopg2==2.9.9
pydantic==1.10.13
pydantic[email]==1.10.13
PyJWT==2.8.0
python-decouple==3.8
requests==2.31.0
SQLAlchemy==1.4.49
uvicorn==0.23.2
